+++ service/spi/src/main/java/gov/va/vro/service/spi/model/GeneratePdfPayload.java
@@ -2,6 +2,7 @@ package gov.va.vro.service.spi.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import gov.va.vro.model.AbdEvidence;
+import gov.va.vro.model.VeteranInfo;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
