--- service/spi/src/main/java/gov/va/vro/service/spi/model/VeteranInfo.java
@@ -1,14 +1,0 @@
-package gov.va.vro.service.spi.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class VeteranInfo {
-  private String first;
-  private String middle;
-  private String last;
-  private String suffix;
-  private String birthdate;
-}
